requirements.txt
rivescript
#PyAudio
SpeechRecognition
PyYaml
phue
vosk
sounddevice
